chore(lint): fix 46 warnings — @ts-ignore→@ts-expect-error, empty interfaces→types, no-unused-expressions, react-refresh disables, split a11y utils
