+++ src/components/generate-form/LyricsVisualEditorCompact.tsx
@@ -39,6 +39,7 @@ import {
 import { SectionCard, EmptyState } from "./lyrics-editor";
 
 // Re-export for backward compatibility with tests
+// eslint-disable-next-line react-refresh/only-export-components
 export { parseLyrics, sectionsToLyrics, sectionsEqual, applyTemplateToSections };
 
 export type { LyricSectionType, LyricSection };
