+++ src/components/music-graph/GraphOnboarding.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 import { useState, useEffect } from "react";
 import { motion, AnimatePresence } from "@/lib/motion";
 import {
