+++ src/components/navigation/RecentlyUsedSection.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 import { useMemo } from "react";
 import { useNavigate } from "react-router";
 import { Clock, ChevronRight } from "@/lib/icons";
