+++ src/components/notifications/smart-alerts/SmartAlertProvider.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 import React, { createContext, useContext, useState, useCallback, useEffect, useRef } from "react";
 import { useNavigate, useLocation } from "react-router";
 import { SmartAlert, MAX_ALERTS_ON_PAGE_LOAD, ALERT_PRIORITIES, QUIET_ROUTES } from "./types";
