+++ src/components/player/PlayerTransitionProvider.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 import { createContext, useContext, useEffect, useId, useMemo, type ReactNode } from "react";
 import { logger } from "@/lib/logger";
