+++ src/components/ui/skeleton.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 import { cn } from "@/lib/utils";
 import { cva, type VariantProps } from "class-variance-authority";
