+++ src/contexts/ThemeContext.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 import { createContext, useContext, useEffect, useState, ReactNode, useCallback } from "react";
 import { logger } from "@/lib/logger";
