+++ src/hooks/telegram/useTelegramSensors.ts
@@ -1,5 +1,3 @@
-// @ts-nocheck - Telegram WebApp API types are dynamic
-
 /**
  * Telegram Sensors Hook
  *
