+++ src/hooks/useAuth.tsx
@@ -1,2 +1,3 @@
 // Re-export everything from AuthContext for backwards compatibility
+/* eslint-disable react-refresh/only-export-components */
 export { useAuth, AuthProvider, type AuthResult } from "@/contexts/AuthContext";
